:tada: refactor store action and add transformIgnorePatterns in package.json
