+++ client/package.json
@@ -186,6 +186,9 @@
     "testMatch": [
       "**/src/**/*.spec.js|**/tests/unit/**/*.spec.(js|jsx|ts|tsx)|**/__tests__/*.(js|jsx|ts|tsx)"
     ],
-    "testURL": "http://localhost/"
+    "testURL": "http://localhost/",
+    "transformIgnorePatterns": [
+      "/node_modules/(?!lodash-es)"
+    ]
   }
 }
